Allow for not dropping indexes when clearing
Indexes may be added through Minidoc::Indexes.ensure_index which runs at the
time the model class is interpreted. Minidoc::TestHelpers.clear_database is
intended to be fired in setup/teardown (or before/after), which would be *after*
the code is loaded and ensure_index has run.

This ordering clears out indexes we intend to be there. If we then have specs
that rely on the indexes being present (maybe we want to test the behavior of
violating an index, or just ensure data-sanity in the specs), these specs break
in non-obvious ways.