Add Minidoc#atomic_set
Minidoc#atomic_set(query, attributes) attempts an update of the record
matching query (along with the instance's _id) and returns true or false
indicating if the update was successful.

The update is said to be atomic because there is no race condition: if
two processes attempt the same update (provided that update changes a
field used in the query), the database guarantees only one will succeed
and return success to its caller.

Example:

    if record.atomic_update({ evaluated_at: nil }, evaluated_at: Time.now)
      # proceed with exactly-once evaluation logic
    else
      # abort, someone else is performing or has performed that logic
    end