Add find_one!
find! was not added because find accepts a selector or an id. In the
case where it accepts an id, find! may make sense, but in the case where
it accepts a selector, the return value is a (truthy) cursor, and the
semantics of find! don't really make snese.

To avoid this confusion, we'll avoid find! entirely. Those users that
what that behavior should use find_one!(_id: x).

Closes #9