Infer association class name when not provided
Defining setters & getters blindly without knowing if a class name was
provided doesn't seem great. Needing to provide a class name explicitly
when in many cases it is easily inferrable also seems odd.

This does a simple inferring of the class name for an association based
on the association name.

I think this simple enough to still be "mini". If we don't want to add
this, though, we probably *shouldn't* define the setter & getter methods
when the `class_name` option isn't provided.